Invert SSR Logic
Inverts the SSR control logic. This makes it
so that when the SSR pin is driven "low", the
controller wants the SSR/heating element "on".

This enables a workaround for low-quality SSRs
which do not truly support switching on 3.3V Pi
logic. This support works by wiring the circuit in
reverse, which produces 2.7V when the circuit is
off (which should not drive the SSR), and 5V when
it is on.