Add pressure reading logic
Adds logic that will read pressure from a
pressurestat attached to an ADS1115 which
is hooked to the Pi's I2C channel. The pressure
conversion function needs some work, but is
close.

Adds a setting to turn this feature on and off.

Adjust the website to allow viewing pressure
information.